pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rouge-ai"
version = "0.0.10"
description = "The SDK that revolutionizes debugging and tracing."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Sudikonda Revanth Kumar"},
]
keywords = ["artificial-intelligence", "agent-systems", "opentelemetry", "tracing", "logging", "debugging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "opentelemetry-api==1.34.1",
    "opentelemetry-sdk==1.34.1",
    "opentelemetry-exporter-otlp==1.34.1",
    "opentelemetry-exporter-otlp-proto-common==1.34.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.34.1",
    "opentelemetry-exporter-otlp-proto-http==1.34.1",
    "opentelemetry-instrumentation==0.55b1",
    "opentelemetry-instrumentation-asgi==0.55b1",
    "opentelemetry-instrumentation-fastapi==0.55b1",
    "opentelemetry-proto==1.34.1",
    "opentelemetry-sdk-extension-aws==2.1.0",
    "opentelemetry-propagator-aws-xray==1.0.2",
    "opentelemetry-semantic-conventions==0.55b1",
    "opentelemetry-util-http==0.55b1",
    "watchtower==3.4.0",
    "pandas>=1.5.0",
    "PyYAML==6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "black==25.1.0",
    "pre-commit==4.2.0",
    "flake8==7.3.0",
    "mypy==1.17.0",
]
fastapi = [
    "fastapi==0.115.12",
    "uvicorn==0.34.3",
    "httpx==0.27.0",
    "opentelemetry-instrumentation-fastapi==0.55b1",
]
llm = [
    "opentelemetry-instrumentation-openai",
    "opentelemetry-instrumentation-anthropic",
    "opentelemetry-instrumentation-cohere",
    "opentelemetry-instrumentation-mistralai",
    "opentelemetry-instrumentation-llamaindex",
    "opentelemetry-instrumentation-haystack",
    "opentelemetry-instrumentation-bedrock",
    "opentelemetry-instrumentation-vertexai",
    "opentelemetry-instrumentation-replicate",
    "opentelemetry-instrumentation-google-generativeai",
    "opentelemetry-instrumentation-langchain",
]
all = [
    "opentelemetry-api==1.34.1",
    "opentelemetry-sdk==1.34.1",
    "opentelemetry-exporter-otlp==1.34.1",
    "opentelemetry-exporter-otlp-proto-common==1.34.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.34.1",
    "opentelemetry-exporter-otlp-proto-http==1.34.1",
    "opentelemetry-instrumentation==0.55b1",
    "opentelemetry-instrumentation-asgi==0.55b1",
    "opentelemetry-instrumentation-fastapi==0.55b1",
    "opentelemetry-proto==1.34.1",
    "opentelemetry-sdk-extension-aws==2.1.0",
    "opentelemetry-propagator-aws-xray==1.0.2",
    "opentelemetry-semantic-conventions==0.55b1",
    "opentelemetry-util-http==0.55b1",
    "watchtower==3.4.0",
    "pandas>=1.5.0",
    "PyYAML==6.0.2",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "black==25.1.0",
    "flake8==7.3.0",
    "mypy==1.17.0",
    "fastapi==0.115.12",
    "uvicorn==0.34.3",
    "pre-commit==4.2.0",
    "pytest==8.4.1",
    "httpx==0.27.0",
    "opentelemetry-instrumentation-fastapi==0.55b1",
    "opentelemetry-instrumentation-openai",
    "opentelemetry-instrumentation-anthropic",
    "opentelemetry-instrumentation-cohere",
    "opentelemetry-instrumentation-mistralai",
    "opentelemetry-instrumentation-llamaindex",
    "opentelemetry-instrumentation-haystack",
    "opentelemetry-instrumentation-bedrock",
    "opentelemetry-instrumentation-vertexai",
    "opentelemetry-instrumentation-replicate",
    "opentelemetry-instrumentation-google-generativeai",
    "opentelemetry-instrumentation-langchain",
]
[project.urls]
Homepage = "https://rouge.ai/"
Documentation = "https://docs.rouge.ai"
Repository = "https://github.com/revanthkumar96/rouge.ai-sdk"
"Bug Reports" = "https://github.com/revanthkumar96/rouge.ai-sdk/issues"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["src"]
include = ["rouge_ai*"]
exclude = ["tests*", "rouge_ai.dashboard.frontend_src*"]
